[Refactor] #32 : Refactor Model, ViewModel, View
