+++ DaeguGoodPriceShop/DaeguGoodPriceShop/Model/ShopAnnotation.swift
@@ -5,38 +5,20 @@
 //  Created by Yeongwoo Kim on 2022/07/05.
 //
 
-import Foundation
 import MapKit
 
-final class ShopAnnotationView: MKAnnotationView {
-    static let identifier = "ShopAnnotationView"
-    
-    override init(annotation: MKAnnotation?, reuseIdentifier: String?) {
-        super.init(annotation: annotation, reuseIdentifier: reuseIdentifier)
-        frame = CGRect(x: 0, y: 0, width: 30, height: 30)
-        centerOffset = CGPoint(x: 0, y: -frame.size.height / 2)
-        setupUI()
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setupUI() {
-        backgroundColor = .clear
-    }
-}
-
 final class ShopAnnotation: NSObject, MKAnnotation {
-    let category: ShopCategory
-    let location: CLLocation
+    let shopSubCategory: ShopSubCategory
+    let latitude: Double
+    let longitude: Double
     var coordinate: CLLocationCoordinate2D {
-        self.location.coordinate
+        CLLocation(latitude: latitude, longitude: longitude).coordinate
     }
     
-    init(category: ShopCategory, location: CLLocation) {
-        self.category = category
-        self.location = location
+    init(shopSubCategory: ShopSubCategory, latitude: Double, longitude: Double) {
+        self.shopSubCategory = shopSubCategory
+        self.latitude = latitude
+        self.longitude = longitude
         super.init()
     }
 }
