+++ DaeguGoodPriceShop/DaeguGoodPriceShop/View/ShopAnnotationView.swift
@@ -1,0 +1,27 @@
+//
+//  ShopAnnotationView.swift
+//  DaeguGoodPriceShop
+//
+//  Created by Shin Jae Ung on 2022/07/06.
+//
+
+import MapKit
+
+final class ShopAnnotationView: MKAnnotationView {
+    static let identifier = "ShopAnnotationView"
+    
+    override init(annotation: MKAnnotation?, reuseIdentifier: String?) {
+        super.init(annotation: annotation, reuseIdentifier: reuseIdentifier)
+        frame = CGRect(x: 0, y: 0, width: 30, height: 30)
+        centerOffset = CGPoint(x: 0, y: -frame.size.height / 2)
+        setupUI()
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func setupUI() {
+        backgroundColor = .clear
+    }
+}
